+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
 
+* 2021-01-04 Support super users in navicore-site.json. Super users can read all documents.
 * 2020-12-22 Add app-protected-link component, to show link address directly in the browser.
 * 2020-12-02 Fix bug: clicking an image should not open another tab. It should open a lightbox to view it.